Revert "Merge branch 'Fix-validation-reset-password' into judelki"
This reverts commit c04f95273338c75c1e39e730a3571ac3da4169f0, reversing
changes made to 26140e556b9038d313cae3a9b2493364e48eff1e.

+++ database/migrations/2023_10_10_223422_create_pagos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('pagos', function (Blueprint $table) {
             $table->id();
             $table->date('fechapago');
-            $table->integer('cantidadpago');
+            $table->float('cantidadpago');
             $table->boolean('estadopago');
             $table->unsignedBigInteger('facturas_id');
             $table->foreign('facturas_id')->references('id')->on('facturas');
